pyproject.toml
[tool.poetry]
package-mode = false
name = "flask-chat"
version = "0.1.0"
description = ""
authors = ["crudenesss <o.bochok26@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
argon2-cffi = "^23.1.0"
flask = "^3.0.3"
flask-socketio = "^5.3.6"
gunicorn = "^22.0.0"
eventlet = "^0.36.1"
wtforms = "^3.1.2"
email-validator = "^2.2.0"
pymongo = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
